# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NSG MOA API Documentation",
        "description": "Nasarawa State Ministry of Agriculture M&E System API documentation",
        "contact": {
            "email": "support@nsgmoa.gov.ng"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/analytics/dashboard": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get dashboard analytics",
                "operationId": "05d184da4e8c54be2eb7790119b8767e",
                "responses": {
                    "200": {
                        "description": "Dashboard analytics data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_farmers": {
                                            "type": "integer"
                                        },
                                        "total_farms": {
                                            "type": "integer"
                                        },
                                        "total_farm_area": {
                                            "type": "number"
                                        },
                                        "active_projects": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/analytics/spatial-distribution": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get spatial distribution of farmers and farms",
                "operationId": "da3a5b8bfee65add8a20f2a647ddefa4",
                "responses": {
                    "200": {
                        "description": "Spatial distribution data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/analytics/weather-trends": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get weather trends and analysis",
                "operationId": "82d5445dd1179058fe38dd8ab37f6283",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Number of days to analyze",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weather trends data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/analytics/farm-productivity": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get farm productivity analysis",
                "operationId": "dc2098a354987b1a1ce438985fd356ae",
                "responses": {
                    "200": {
                        "description": "Farm productivity data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}